Applied all the suggestions.
Please review then approve if there are no further suggestions for the MVC.
