+++ models/users.json
@@ -7,7 +7,7 @@
     "firstName": "Axel",
     "lastName": "Alba",
     "profilePic": "https://cdn.shopify.com/s/files/1/2580/0020/files/TontonHouse-______180221_6ef85f36-b495-4e0b-8ff7-752c2402c478_300x.png?v=1529488749",
-    "bio": "I am Tonton and I have friends.",
+    "bio": "Cute ko daw kasi ako si tonton",
     "mobile": "09236457121",
     "telephone": "4495967",
     "address": "Gotham City, Secret",
