+++ views/gallerySample.hbs
@@ -1,4 +1,4 @@
-
+<script src="/js/gallery.js"></script>
 
 <div class = "gallery"> <!-- just insert this div block to access the gallery module--> 
     <!-- This part would be implemented in the script.js, please implement post.hbs
