+++ views/layouts/main.hbs
@@ -15,8 +15,6 @@
         <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
         <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
 
-        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script> <!--This is needed for other features to work (safe for editProfile)-->
-
 
     <!-- css for fancyBox -->
         <link rel = "stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/fancybox/3.5.7/jquery.fancybox.min.css">
@@ -27,13 +25,6 @@
         <link rel="stylesheet" href="/css/profilePage.css">
         <link rel="stylesheet" href="/css/newsfeed.css"> <!-- newsfeed.hbs css-->
         <link rel="stylesheet" href="/css/posts.css"> <!-- post.hbs css--> 
-
-        <!--Script for the Ajax-->
-        <script src="/js/script.js"></script>
-
-
-        <!--Scripts for features-->
-        <script src="/js/editProfile.js"></script>
     </head>
     <body>
         {{{body}}}
