+++ views/post.hbs
@@ -1,3 +1,4 @@
+<script src="/js/gallery.js"></script>
 <script src="/js/post.js"></script>
 <input type="hidden" id="pid" value={{pid}}>
 <div class="row">
