+++ views/profile.hbs
@@ -1,9 +1,10 @@
+<script src="/js/gallery.js"></script>
 <!--This is where the navbar should be-->
 <!--Profile picture & biography-->
 <div id = "profileHead" class = "container-fluid padding">
 	<div id = "profileHeader" class = "row welcome text-center fixed-wrap">
         <div class = "col-12">
-            <a href = "editProfile.hbs"> <img src= "img/settings.png" style = "position: absolute; 
+            <a href = "/editProfile"> <img src= "img/settings.png" style = "position: absolute; 
             top: 25px;
             right: 25px;
             height: auto;
